+++ Documentation/Source/devEditor.Target.cs
@@ -6,7 +6,9 @@ public class devEditorTarget : TargetRules
     public devEditorTarget( TargetInfo Target) : base(Target)
     {
         Type = TargetType.Editor;
-#if UE_5_4_OR_LATER
+#if UE_5_7_OR_LATER
+        DefaultBuildSettings = BuildSettingsVersion.V6;
+#elif UE_5_4_OR_LATER
         DefaultBuildSettings = BuildSettingsVersion.V4;
 #else
         DefaultBuildSettings = BuildSettingsVersion.V2;