+++ Source/CesiumRuntime/CesiumRuntime.Build.cs
@@ -105,6 +105,7 @@ public class CesiumRuntime : ModuleRules
                 "Slate",
                 "SlateCore",
                 "Niagara"
+                "ChaosCore"
             }
         );
 
@@ -157,8 +158,13 @@ public class CesiumRuntime : ModuleRules
             }
         );
 
-        ShadowVariableWarningLevel = WarningLevel.Off;
+#if UE_5_7_OR_LATER
+        IncludeOrderVersion = EngineIncludeOrderVersion.Unreal5_7;
+        CppCompileWarningSettings.ShadowVariableWarningLevel = WarningLevel.Off;
+#else
         IncludeOrderVersion = EngineIncludeOrderVersion.Unreal5_4;
+        ShadowVariableWarningLevel = WarningLevel.Off;
+#endif
         PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
 
         CppStandard = CppStandardVersion.Cpp20;