+++ Source/CesiumRuntime/Private/CesiumPrimitive.h
@@ -4,6 +4,7 @@
 
 #include "Cesium3DTileset.h"
 #include "CesiumEncodedMetadataUtility.h"
+#include "CesiumLoadedTile.h"
 #include "CesiumMetadataPrimitive.h"
 #include "CesiumPrimitiveFeatures.h"
 #include "CesiumPrimitiveMetadata.h"
@@ -122,7 +123,7 @@ public:
 };
 
 UINTERFACE()
-class UCesiumPrimitive : public UInterface {
+class UCesiumPrimitive : public UCesiumLoadedTilePrimitive {
   GENERATED_BODY()
 };
 
@@ -135,11 +136,17 @@ class UCesiumPrimitive : public UInterface {
  * code reuse and make certain functions (e.g., UpdateTransformFromCesium())
  * simpler.
  */
-class ICesiumPrimitive {
+class ICesiumPrimitive : public ICesiumLoadedTilePrimitive {
   GENERATED_BODY()
 public:
   virtual CesiumPrimitiveData& getPrimitiveData() = 0;
   virtual const CesiumPrimitiveData& getPrimitiveData() const = 0;
+
+  // from ICesiumLoadedTilePrimitive:
+  const CesiumGltf::MeshPrimitive* GetMeshPrimitive() const override;
+  const FCesiumPrimitiveFeatures& GetPrimitiveFeatures() const override;
+  const FCesiumPrimitiveMetadata& GetPrimitiveMetadata() const override;
+
   virtual void
   UpdateTransformFromCesium(const glm::dmat4& CesiumToUnrealTransform) = 0;
 };