+++ CHANGES.md
@@ -4,7 +4,27 @@
 
 ##### Fixes :wrench:
 
+- Fixed a bug that prevented `UCesiumPrimitiveFeaturesBlueprintLibrary::GetPrimitiveFeatures` from retrieving the features of instanced meshes.
+
+### v2.21.0 Preview for Unreal Engine 5.7 - 2025-11-17
+
+##### Additions :tada:
+
+- Added support for Unreal Engine 5.7.
+
+### v2.21.0 - 2025-11-03
+
+##### Additions :tada:
+
+- Added `CesiumAzureMapsRasterOverlay`.
+- Added the interface `ICesium3DTilesetLifecycleEventReceiver`: when an implementation is registered on a tileset (with `ACesium3DTileset::SetLifecycleEventReceiver`), its functions will be called at various points in a tile's lifecycle, like when a mesh component is created, when a material is instanced, when the tile changes visibility, when it is unloaded, etc.
+
+##### Fixes :wrench:
+
+- Fixed a bug where `CesiumCreditSystem` would not filter out empty credits, resulting in duplicate on-screen delimiters.
 - Fixed a problem where multi-selecting `UCesiumGlobeAnchorComponent` could cause the selected components to teleport to 0 degrees longitude and 0 degrees latitude. Now, the geospatial position and orientation fields are hidden while multi-selecting.
+
+In addition to the above, this release updates [cesium-native](https://github.com/CesiumGS/cesium-native) from v0.52.0 to v0.53.0. See the [changelog](https://github.com/CesiumGS/cesium-native/blob/main/CHANGES.md) for a complete list of changes in cesium-native.
 
 ### v2.20.0 - 2025-10-01